+++ nn_dataflow/core/scheduling.py
@@ -19,6 +19,7 @@ program. If not, see <https://opensource.org/licenses/BSD-3-Clause>.
 """
 
 from collections import OrderedDict, namedtuple
+import math
 import fastcache
 
 from . import data_category_enum as de
@@ -267,24 +268,34 @@ class Scheduling(object):
         '''
         Make the schedule result from loop blocking and partitioning.
         '''
-        total_nhops = [unh * f for unh, f
-                       in zip(unit_nhops, lbs.get_top_level_fetch())]
-        cost_loop = lbs.get_cost(self.cost)
-        cost_part = self.cost.noc_hop * sum(total_nhops)
-
         scheme = OrderedDict()
 
+        # Cost components.
+        cost_access = lbs.get_access_cost(self.cost)
+
+        total_nhops = [unh * f for unh, f
+                       in zip(unit_nhops, lbs.get_top_level_fetch())]
+        cost_noc = self.cost.noc_hop * sum(total_nhops)
+
+        cost_op = self.cost.mac_op * lbs.ops
+
+        cost_static = self.cost.unit_static * lbs.time * lbs.num_nodes
+
+        assert not math.isnan(cost_op + cost_access + cost_noc + cost_static)
+
         # Overall stats.
-        scheme['cost'] = cost_loop + cost_part
+        scheme['cost'] = cost_op + cost_access + cost_noc + cost_static
         scheme['time'] = lbs.time
         scheme['ops'] = lbs.ops
-        scheme['num_nodes'] = part.size()
-        scheme['cost_loop'] = cost_loop
-        scheme['cost_part'] = cost_part
+        scheme['num_nodes'] = lbs.num_nodes
+        scheme['cost_op'] = cost_op
+        scheme['cost_access'] = cost_access
+        scheme['cost_noc'] = cost_noc
+        scheme['cost_static'] = cost_static
         scheme['proc_time'] = lbs.proc_time
         scheme['bus_time'] = lbs.bus_time
         scheme['dram_time'] = lbs.dram_time
-        scheme['access'] = lbs.access
+        scheme['access'] = lbs.get_access()
         scheme['total_nhops'] = total_nhops
         scheme['fetch'] = lbs.fetch